cfx_execute_helper::observer

Module exec_tracer

Modules

action_types

filter

trace_types

Structs

BlockExecTraces

Represents all traces produced by transactions in a single block.

Call

Description of a call action, either a CALL operation or a message transaction.

CallResult

Description of the result of a call action.

Create

Description of a create action, either a CREATE operation or a create transaction.

CreateResult

Description of the result of a create action.

ErrorUnwind

An executive tracer only records errors during EVM unwind.

ExecTrace

Trace localized in vector of traces produced by a single transaction.

ExecTraceKey

ExecTracer

Simple executive tracer. Traces all calls and creates.

InternalTransferAction

Description of the result of an internal transfer action regarding about CFX.

LocalizedTrace

SelfDestructAction

Represents a selfdestruct action fka suicide.

SetAuth

TraceFilter

Log event Filter.

TransactionExecTraces

Represents all traces produced by a single transaction.

Enums

Action

Description of an action that we trace; will be either a call or a create.

ActionType

Auto-generated discriminant enum variants

Outcome

The outcome of the action result.

Functions

recover_phantom_trace_for_call

recover_phantom_trace_for_withdraw

recover_phantom_traces